feat: v2 configuration support
