+++ src/task/cluster/test.rs
@@ -1,4 +1,4 @@
-use crate::task::ins::Type;
+use crate::task::cluster::ins::Type;
 
 use super::*;
 
@@ -14,15 +14,15 @@ fn test_cluster_sorted_after_from() {
     println!(
         "test_cluster_sorted_after_from: {:?}",
         cluster
-            .instances
+            .topology()
             .iter()
             .map(|instance| &instance.itype)
             .collect::<Vec<&Type>>()
     );
 
-    assert_eq!(&cluster.instances[0].itype, &Type::Router);
-    assert_eq!(&cluster.instances[1].itype, &Type::Storage);
-    assert_eq!(&cluster.instances[2].itype, &Type::Storage);
-    assert_eq!(&cluster.instances[3].itype, &Type::Custom);
-    assert_eq!(&cluster.instances[4].itype, &Type::Custom);
+    assert_eq!(&cluster.topology()[0].itype, &Type::Router);
+    assert_eq!(&cluster.topology()[1].itype, &Type::Storage);
+    assert_eq!(&cluster.topology()[2].itype, &Type::Storage);
+    assert_eq!(&cluster.topology()[3].itype, &Type::Custom);
+    assert_eq!(&cluster.topology()[4].itype, &Type::Custom);
 }
